requirements.txt
certifi
charset-normalizer
click
decorator
idna
jsonschema
gradio==3.0.0
keras
matplotlib
matplotlib-inline
networkx==2.6.3
numpy
openapi-python-client==0.13.4
packaging
pandas
Pillow
protobuf==3.20.3
pydantic==1.10.2
pyparsing
py-solc-x
python-dateutil
pytz
requests
scipy
six
Solc
solidity-parser==0.0.4
spacy
spektral==1.1.0
tenacity
tensorflow==2.12.0
termcolor
typer==0.7.0
urllib3
wcwidth
antlr4-python3-runtime==4.8